src/com.elektrobit.ebrace.core.datamanager/src/com/elektrobit/ebrace/core/datamanager/internal/runtime/event/MinMax.java
/*******************************************************************************
 * Copyright (C) 2018 Elektrobit Automotive GmbH
 * 
 * This program and the accompanying materials are made
 * available under the terms of the Eclipse Public License 2.0
 * which is available at https://www.eclipse.org/legal/epl-2.0/
 * 
 * SPDX-License-Identifier: EPL-2.0
 ******************************************************************************/
package com.elektrobit.ebrace.core.datamanager.internal.runtime.event;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class MinMax<T>
{
    private T min;
    private T max;

    public boolean contains(T value)
    {
        boolean equalsMin = value.equals( min );
        boolean equalsMax = value.equals( max );
        return equalsMin || equalsMax;
    }
}
